Supporting autocomplete promises
When the user types into the tagpicker input field, the autocomplete
function is called. The autocomplete function is called with two
arguments, the user's current input and a results object.

The autocomplete function should set results.data to be the appropriate
array of tags given the user's current input. For example,

      function foo(search, results) {
        results.data = [search + ' foo', search + ' bar'];
      }

When the user types 'hello' into the tagpicker's input, the tagpicker
results will render the array ['hello foo', 'hello bar'].